+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 ## Changelog
 
 - [command_toolbox.apply_velocity] feature: new module to provide `/apply-velocity` command.
+- [command_toolbox.{burn/freeze}] feature: `/burn` and `/freeze` commands now support to select a collection of entities as target.